feat: added warping dtw before we do the tune-following?
Reread the paper and noticed:

"
Basically, for any query-template pair of pitch vectors be-
ing compared, the optimal warping path is found with the
DTW algorithm. Then, the consecutive pitch values qi of
the time-warped query are compared to the corresponding
pitch values ti in the template, in order to produce a mod-
ified query sequence ˆq.
"

I think this was the right impl? didnt change much